Refactor(performance): Split the metrics reporter
reporter.py held 853 lines, past the 400-line reviewability limit the
aislop code-quality gate applies.

Split it into a package at the same dotted path: models (the metric
and alert enums, records and the report), collector (MetricsCollector
accumulating bounded per-name history), visualizer (the ASCII charts
and the HTML export template) and reporter (PerformanceReporter wiring
the two, computing trends and evaluating alerts). The package __init__
re-exports the previous names.
